pyproject.toml
[tool.poetry]
name = "mc-playerdb"
version = "2.0.0"
description = "A minecraft datapack which aids in storing data per player"
authors = ["Ritik Shah <ritik10.shah@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
colour = "^0.1.5"
beet = "^0.40.1"
lectern = "^0.15.3"

[tool.beet]
output = "build"
extend = ["beet-project.yaml"]

[[tool.beet.pipeline]]
output = "build"
extend = ["beet-project.yaml"]
data_pack.zipped = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
